backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-cloud-firestore==2.16.0
google-cloud-pubsub==2.21.1
python-dotenv==1.0.0
pydantic>=2.0,<3.0
httpx>=0.27,<1.0
cachetools>=5.3,<6.0
